requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
requests>=2.28.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.10.0

# OCR and Image Processing
pytesseract>=0.3.8
opencv-python>=4.5.0
Pillow>=9.0.0
pdf2image>=1.16.0
PyMuPDF>=1.20.0

# Scientific Computing
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
shapely>=1.8.0
networkx>=3.0.0

# Caching and Performance
redis>=4.0.0
celery>=5.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.24.0,<0.28.0

# Template Engine and Reporting
jinja2>=3.0.0
weasyprint>=68.0

# Authentication and Security
PyJWT>=2.6.0
python-jose[cryptography]>=3.3.0

# Development
python-dotenv>=0.19.0

# System Monitoring
psutil>=5.9.0
